fix: 정답을 본 풀이인 경우 풀이 제출이 시간이 없을때 안되는 버그 fix

+++ src/main/java/kr/solta/adapter/webapi/response/SolvedRegisterResponse.java
@@ -6,7 +6,7 @@ import kr.solta.domain.Tier;
 
 public record SolvedRegisterResponse(
         long solvedId,
-        long solveTimeSeconds,
+        Integer solveTimeSeconds,
         ProblemDetail problem
 ) {
     record ProblemDetail(
@@ -18,7 +18,7 @@ public record SolvedRegisterResponse(
 
     public static SolvedRegisterResponse from(Solved solved) {
         Problem problem = solved.getProblem();
-        
+
         return new SolvedRegisterResponse(
                 solved.getId(),
                 solved.getSolveTimeSeconds(),
